Add true light status reporting and fix a couple minor bugs.
FPP will now monitor the state of the Pixel Overlay Model behind a
virtual light and send a light state update message when the on/off
status or color changes.  The current color is read from the first
pixel in a model is used for the color.  These state checks occur
once per second.  Previously the incoming command handler would just
return a message indicating the desired state was now the current state,
but now the state monitor returns the true state.  This allows the model
to be controlled outside of HA and the state changes will be reflected
inside HA.